+++ packages/ui/src/form/primitives/select/popup.tsx
@@ -21,7 +21,7 @@ export const Popup = forwardRef< HTMLDivElement, SelectPopupProps >(
 		ref
 	) {
 		const popupContent = (
-			<ThemeProvider>
+			<ThemeProvider crossesPortalBoundary>
 				<_Select.Popup
 					ref={ ref }
 					className={ clsx( itemPopupStyles.popup, className ) }